More hacks to find out why I'm not getting sensible readings in the cluster

+++ nbresuse/config.py
@@ -115,7 +115,7 @@ class ResourceUseDisplay(Configurable):
     ).tag(config=True)
 
     disk_limit = Union(
-        trait_types=[Float(), Callable()],
+        trait_types=[Int(), Callable()],
         default_value=0,
         help="""
         Disk usage limit to display to the user.
